Split up tests and rename the dfs
